#101 rename: 테스트 클래스 이동
